+++ packages/examples/src/views/custom-buttons/index.tsx
@@ -3,8 +3,8 @@ import { useSearchParams } from "react-router-dom";
 import {AiOutlineZoomIn, AiOutlineZoomOut,AiOutlineFullscreenExit, AiOutlineFullscreen, AiFillPlayCircle, AiFillPauseCircle} from "react-icons/ai";
 import {MdFilterCenterFocus} from "react-icons/md";
 
-import { SigmaContainer, ControlsContainer, ZoomControl, SearchControl , FullScreenControl} from "@react-sigma/core";
-import {LayoutForceAtlas2Control} from "@react-sigma/layout-forceatlas2";
+import { SigmaContainer, ControlsContainer, ZoomControl, FullScreenControl} from "@react-sigma/core";
+import { LayoutForceAtlas2Control } from "@react-sigma/layout-forceatlas2";
 import { MyGraph } from "./my-graph";
 
 export const CustomButtonsView: FC = () => {