+++ packages/examples/src/views/home.tsx
@@ -4,27 +4,23 @@ import { Link } from "react-router-dom";
 export const HomeView: FC = () => {
   return (
     <div className="container-fluid">
-      <h1>@react-sigma - examples</h1>
+      <h1>List of examples</h1>
       <ul>
         <li>
-          <Link to="/complete" title="Complete">
-            Complete
-          </Link>
+          <Link to="/complete" title="Complete">Complete</Link>{" "}
+           A complete tour of the lib with all available features
         </li>
         <li>
-          <Link to="/events" title="Events">
-            Events
-          </Link>
+          <Link to="/events" title="Events">Events</Link>{" "}
+          Open the js console and see all the tracked events
         </li>
         <li>
-          <Link to="/multiple" title="Multiple">
-            Multiples
-          </Link>
+          <Link to="/multiple" title="Multiple">Multiples</Link>{" "}
+          Display multiple independant graph on the same page
         </li>
         <li>
-          <Link to="/custom-buttons" title="Custom Buttons">
-            Custom Buttons
-          </Link>
+          <Link to="/custom-buttons" title="Custom Buttons">Custom Buttons</Link>{" "}
+          Make your own style by overring the native display of controllers
         </li>
       </ul>
     </div>