+++ packages/website/docusaurus.config.js
@@ -119,7 +119,6 @@ const config = {
       },
     }),
   plugins: [
-    ["docusaurus-plugin-sass", {}],
     [
       "docusaurus-plugin-typedoc",